// src/api/back/models/RecordType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum RecordType {
    KILLS = 'KILLS',
    KDA = 'KDA',
    ASSISTS = 'ASSISTS',
    DEATHS = 'DEATHS',
    LASTHITS = 'LAST_HITS',
    DENIES = 'DENIES',
    GPM = 'GPM',
    XPM = 'XPM',
    NETWORTH = 'NETWORTH',
    TOWERDAMAGE = 'TOWER_DAMAGE',
    HERODAMAGE = 'HERO_DAMAGE',
    HEROHEALING = 'HERO_HEALING'
}

export function RecordTypeFromJSON(json: any): RecordType {
    return RecordTypeFromJSONTyped(json, false);
}

export function RecordTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecordType {
    return json as RecordType;
}

export function RecordTypeToJSON(value?: RecordType | null): any {
    return value as any;
}